eddie: Add hinting API to share data
This change adds a new helper API that allows contract instances or other
structs to be associated with elements of a program.  All configured contracts
are available as hints at the outset, which simplifies the process of making
use of already-annotated members.

The ReturnConcrete contract has been updated to use this mechanism and we see
in the test code that more functions are provably clean, since we short-circuit
when examining calls to similarly-annotated functions.